Create subnetting of IpAdds.cpp
The program starts by including the input/output library and the standard library. It then prompts the user to enter the IP address and the subnet mask, which are stored as integers. The number of subnet bits and host bits are calculated by using bitwise operations. Finally, the program displays the number of subnet bits and host bits to the user.

This is a basic example, and there are many other ways to perform subnetting using C++, and many other things you can do with the subnet information, like generate the IPs of the subnet, and many other features.